// src/enums/Type.ts
export enum Type {
    INODE = 0x0000,
    NULL = 0x0001,
    STRUCTURE = 0x0002,
    ARRAY = 0x0004,
    STRING = 0x0008,
    WSTRING = 0x0010,
    BOOL = 0x0020,
    CHAR = 0x0040,
    WCHAR = 0x0080,
    SHORT = 0x0100,
    INTEGER = 0x0200,
    FLOAT = 0x0400,
    LONG = 0x0800,
    DOUBLE = 0x1000
}